requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
black==25.1.0
click==8.2.1
dj-database-url==3.0.0
Django==5.2.3
django-bootstrap4==25.1
django-filter==25.1
djangorestframework==3.16.0
isort==6.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
soupsieve==2.7
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.0
